requirements.txt
certifi>=2022.6.15
charset-normalizer>=2.1.0
idna>=3.3
numpy>=1.19.5
panda>=0.3.1
pandas>=1.1.5
python-dateutil>=2.8.2
pytz>=2022.1
requests>=2.27.1
six>=1.16.0
urllib3>=1.26.11
PyYAML>=6.0
smmap>=5.0.0
gitdb>=4.0.9
GitPython>=3.1.20
